+++ freecad/AIGenFurniture/manufacturing/__init__.py
@@ -8,56 +8,67 @@
 EXPORT_DEFINITIONS_SCHEMA_VERSION = "1.0"
 
 EXPORT_DEFINITIONS = {
+    # this export is shop-specific for my shop. Should not be part of any release
     "export_pal_for_proficut":{
         "enabled": False, #False for MVP
         "runner": "export_pal_for_proficut",
         "module": "export_for_proficut",
     },
+    # this export is shop-specific for my shop. Should not be part of any release
     "export_pfl_for_proficut": {
         "enabled": False, #False for MVP
         "runner": "export_pfl_for_proficut",
         "module": "export_for_proficut",
     },
+    # this export is shop-specific for my shop. Should not be part of any release
     "export_front_for_nettfront": {
         "enabled": False, #False for MVP
         "runner": "export_front_for_nettfront",
         "module": "export_for_nettfront",
     },
+    # this export is generic. It outputs all elements that inherit Board class,
     "export_csv": {
         "enabled": True, #True for MVP
         "runner": "export_csv",
         "module": "export_csv",
     },
+    # this export is generic for all projects
     "export_stl": {
         "enabled": True, #False for MVP
         "runner": "export_stl_order",
         "module": "export_stl_new",
     },
+    # this export is generic for all elements
     "export_cost_sheet":{
         "enabled": False, #False for MVP
         "runner": "export_cost_sheet",
         "module": "generate_offer_cost",
     },
+    # this is shop specific - each shop will want a different type of print
     "print_order_summary":{
         "enabled": False, #False for MVP
         "runner": "print_order_summary",
         "module": "generate_offer_cost",
     },
+    # this is shop specific - each shop will want a different offer file format
     "generate_offer_file":{
         "enabled": False, #False for MVP
         "runner": "generate_offer_file",
         "module": "generate_offer_cost",
     },
+    # this is shop specific - each shop will want a different offer file format
     "generate_assembly_file":{
         "enabled": False, #False for MVP
         "runner": "generate_assembly_file",
         "module": "generate_assembly_file",
     },
+    # this is shop specific - each shop will want a different offer file format
     "generate_drill_file":{
         "enabled": False, #False for MVP
         "runner": "generate_drill_file",
         "module": "generate_assembly_file",
     },
+    # this is shop specific - each shop will want a different offer file format
     "generate_drill_pdf_reportlab":{
         "enabled": False, #False for MVP
         "runner": "generate_drill_pdf_reportlab",